* develop: (89 commits)
  Fix integration tests
  Adapt webclient-backend to the latest changes on develop
  Update license year
  Move tempCompressedDataFile deletion to XoClientService
  Rename cachedDataFile to tempCompressedDataFile
  Fix talk tool
  Move SelectedFile to common
  Format code
  Delete cached files if user decides to not send the attachments
  Fix wrong constant
  Use getContentLength instead of calculating it
  Use filePath for computeHmac signature
  Inline fileLength
  Use constants to check mediaType
  Check if video file exists on select
  Refactor SelectedContent and its derived classes for clarity
  Raise TalkServer version to 3.0.32
  Fix static routing on server
  Rename contentType to mimeType and contentMediaType to mediaType
  Raise TalkServer version to 3.0.31
  ...

Conflicts:
	android/src/main/java/com/hoccer/xo/android/content/selector/AudioSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/CaptureSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/ContactSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/DataSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/ImageSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/LocationSelector.java
	android/src/main/java/com/hoccer/xo/android/content/selector/VideoSelector.java
	android/src/main/java/com/hoccer/xo/android/view/chat/ChatMessageItem.java
	android/src/main/java/com/hoccer/xo/android/view/chat/attachments/ChatAudioItem.java
	android/src/main/java/com/hoccer/xo/android/view/chat/attachments/ChatContactItem.java

+++ client/src/main/java/com/hoccer/talk/client/IXoTransferListener.java
@@ -7,5 +7,5 @@ public interface IXoTransferListener extends IProgressListener {
 
     public void onStateChanged(IXoTransferState state);
 
-    public void onProgressUpdated(int progress, int contentLength);
+    public void onProgressUpdated(long progress, long contentLength);
 }
